refactor(stable_api): extract AssetsMixin; stable_api becomes facade
Final mixin extraction. Moves 7 asset metadata/payout methods
(get_instruments, get_all_asset_name, get_available_asset,
check_asset_open, get_all_assets, get_payment, get_payout_by_asset)
into the AssetsMixin. After this commit, stable_api.py is a thin
facade containing only Quotex.__init__, websocket, set_session,
check_connect, _check_connect, close, and re_subscribe_stream;
all domain methods live in pyquotex/_api/*. Public surface unchanged.
